refactor: move getRottXComponent to a util
